feat: Minor Refactoring
